* Bugfix for int type modifications (short -> int16_t instead of int), breaks PPC backup and restore
svn path=/trunk/flamerobin/; revision=632

+++ src/ibpp/_spb.cpp
@@ -71,7 +71,7 @@ void SPB::Insert(char opcode)
 
 void SPB::InsertString(char type, int lenwidth, const char* data)
 {
-	int len = (int)strlen(data);
+	int16_t len = (int16_t)strlen(data);
 
 	Grow(1 + lenwidth + len);
 	mBuffer[mSize++] = type;
